printf: some size_t to long int

+++ srcs/ft_printf.c
@@ -26,9 +26,9 @@ static void	init_store(t_store *store)
 
 int	ft_printf(char const *s, ...)
 {
-	va_list	arg;
-	size_t	nb_chars;
-	t_store	store[NB_OF_CONV + 1];
+	va_list		arg;
+	long int	nb_chars;
+	t_store		store[NB_OF_CONV + 1];
 	int		i;
 
 	nb_chars = 0;
